requirements.txt
dash==1.16.3
dash-bootstrap-components==0.10.7
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
flask-compress==1.5.0
pandas==1.1.3
numpy==1.19.2
flask-login==0.5.0
flask-sqlalchemy==2.4.3
flask==1.1.2
gunicorn
plotly==4.11.0
plotly-express==0.4.0
sqlalchemy==1.3.19
werkzeug==1.0.1
psycopg2==2.8.6
Pillow==8.0.1
requests==2.25.0
Flask-Caching==1.9.0
google-cloud-storage==1.33.0
google-auth==1.23.0
